extract argparser to own file
